Reorganize files internally
