+++ fooof/core/funcs.py
