+++ fooof/core/reports.py
@@ -5,7 +5,7 @@ import os
 import matplotlib.pyplot as plt
 from matplotlib import gridspec
 
-from fooof.strings import gen_settings_str, gen_results_str_fm, gen_results_str_fg
+from fooof.core.strings import gen_settings_str, gen_results_str_fm, gen_results_str_fg
 from fooof.plts.fg import plot_fg_bg, plot_fg_gf, plot_fg_osc_cens
 
 ###################################################################################################
