+++ fooof/core/strings.py
