+++ fooof/fit.py
@@ -7,14 +7,12 @@ import numpy as np
 from scipy.optimize import curve_fit
 
 from fooof.utils import trim_psd, mk_freq_vector
-from fooof.utils import group_three, check_array_dim
-from fooof.utils import get_attribute_names, docs_drop_param
-
-from fooof.io import save_fm, load_json
 from fooof.plts.fm import plot_fm
-from fooof.reports import create_report_fm
-from fooof.strings import gen_settings_str, gen_results_str_fm, gen_report_str, gen_bw_warn_str
-from fooof.funcs import gaussian_function, expo_function, expo_nk_function
+from fooof.core.io import save_fm, load_json
+from fooof.core.reports import create_report_fm
+from fooof.core.funcs import gaussian_function, expo_function, expo_nk_function
+from fooof.core.utils import group_three, check_array_dim, get_attribute_names, docs_drop_param
+from fooof.core.strings import gen_settings_str, gen_results_str_fm, gen_report_str, gen_bw_warn_str
 
 ###################################################################################################
 ###################################################################################################
