+++ fooof/tests/test_analysis.py
