+++ fooof/tests/test_funcs.py
@@ -3,7 +3,7 @@
 import numpy as np
 from scipy.stats import norm, linregress
 
-from fooof.funcs import *
+from fooof.core.funcs import *
 
 ###################################################################################################
 ###################################################################################################
