+++ fooof/tests/test_utils.py
@@ -4,6 +4,7 @@ from py.test import raises
 import numpy as np
 
 from fooof.utils import *
+from fooof.core.utils import *
 
 ###################################################################################################
 ###################################################################################################
